+++ src/crab/ebpf_checker.cpp
@@ -7,7 +7,6 @@
 #include <optional>
 #include <utility>
 
-#include "asm_ostream.hpp"
 #include "asm_unmarshal.hpp"
 #include "config.hpp"
 #include "crab/array_domain.hpp"
