+++ src/crab/ebpf_domain.cpp
@@ -9,7 +9,6 @@
 
 #include "boost/endian/conversion.hpp"
 
-#include "asm_ostream.hpp"
 #include "asm_unmarshal.hpp"
 #include "config.hpp"
 #include "crab/array_domain.hpp"
