+++ src/crab/ebpf_transformer.cpp
@@ -10,7 +10,6 @@
 
 #include "boost/endian/conversion.hpp"
 
-#include "asm_ostream.hpp"
 #include "asm_unmarshal.hpp"
 #include "config.hpp"
 #include "crab/array_domain.hpp"
